refactor(composer): move config resolution from component to service
To unload droppable component
